Fixed Complimentary Filter Calculation
The filter that was previously being applied weighted the integral of the gyro data. This lead to a huge drift in the angle calculation of the quadcopter. This commit fixes this by using the instantaneous change in angle (provided by the groll/gpitch * dt term) as a majority weight in estimating the angle of the quadcopter.